requirements.txt
streamlit==1.41.1
google-generativeai==0.4.0
python-dotenv==1.0.1
